Add info on number of currently displayed aircraft
Adds a small function to list the number of aircraft displayed in the
web based UI. This counts both actively tracked and 'stale' aircraft.

This is useful in cases where the LingerTime is set very high and the
user wishes to see how many aircraft were observed in X amount of time.